관리부 수정

+++ geumgang/src/main/java/Model/GwanriDTO/AgreementConditionDTO.java
@@ -4,16 +4,16 @@ import java.io.Serializable;
 
 @SuppressWarnings("serial")
 public class AgreementConditionDTO implements Serializable { //협약조건안DTO
-	Long agreementConditionNum; //협약조건 번호
+	Integer agreementConditionNum; //협약조건 번호
 	String agreementConditionSubject; //협약조건 제목
 	java.sql.Date agreementConditionDate; //협약조건 작성일
 	float agreementConditionRatio; //업체희망정산비율
 	Integer agreementConditionSett; //정산기준시간
 	String gwanRiNum; //관리부직원번호
-	public Long getAgreementConditionNum() {
+	public Integer getAgreementConditionNum() {
 		return agreementConditionNum;
 	}
-	public void setAgreementConditionNum(Long agreementConditionNum) {
+	public void setAgreementConditionNum(Integer agreementConditionNum) {
 		this.agreementConditionNum = agreementConditionNum;
 	}
 	public String getAgreementConditionSubject() {
